requirements.txt
absl-py==0.11.0
astunparse==1.6.3
auditok==0.2.0
cached-property==1.5.2
cachetools==4.2.1
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
cycler==0.10.0
ffmpeg==1.4
ffpyplayer==4.3.2
filelock==3.0.12
flatbuffers==1.12
gast==0.3.3
google-auth==1.27.0
google-auth-oauthlib==0.4.2
google-pasta==0.2.0
grpcio==1.32.0
h5py==2.10.0
huggingface-hub==0.0.7
HyperPyYAML==1.0.0
idna==2.10
importlib-metadata==3.7.0
joblib==1.0.1
Keras==2.4.3
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
keras-vggface==0.6
kiwisolver==1.3.1
Markdown==3.3.4
matplotlib==3.3.4
mtcnn==0.1.0
numpy==1.19.5
oauthlib==3.1.0
opencv-python==4.5.1.48
opt-einsum==3.3.0
packaging==20.9
Pillow==8.1.1
progressbar2==3.53.1
protobuf==3.15.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pydload==1.0.9
pyparsing==2.4.7
python-dateutil==2.8.1
python-utils==2.5.6
PyYAML==5.4.1
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
ruamel.yaml==0.16.13
ruamel.yaml.clib==0.2.2
scikit-learn==0.24.1
scipy==1.5.4
sentencepiece==0.1.95
seqeval==0.0.3
six==1.15.0
sounddevice==0.4.1
SoundFile==0.10.3.post1
speechbrain==0.5.4
SpeechRecognition==3.8.1
tensorboard==2.4.1
tensorboard-plugin-wit==1.8.0
tensorflow-estimator==2.4.0
tensorflow-gpu==2.4.1
termcolor==1.1.0
threadpoolctl==2.1.0
torch==1.8.0
torchaudio==0.8.0
tqdm==4.59.0
typing-extensions==3.7.4.3
urllib3==1.26.3
Werkzeug==1.0.1
wget==3.2
wrapt==1.12.1
zipp==3.4.0
